feat: list all user's profiles

+++ routes/profiles.js
@@ -4,6 +4,9 @@ module.exports = router;
 
 const db = require('../db');
 
+const auth = require('../authMW');
+
+
 //TODO: add the /profiles endpoint that returns all the user's profiles.
 // this requires euthentication - only works when the user is logged in
 
@@ -15,6 +18,23 @@ const db = require('../db');
 //         DEVICE, DEVICE //NOTE: see ./devices.js for syntax
 //     ]
 // }
+
+//[auth]
+///get all the user's profiles
+router.get('/', auth, async (req, res) => {
+
+    if (!req.id_user) {
+        res.status(401).send("You must be logged in to list all the profiles.");
+        return;
+    }
+
+    let profiles = await db.getProfiles(req.id_user);
+
+
+    //TODO: maybe 200 isn't the right code? idk 
+    res.status(200).send(JSON.stringify(profiles));
+});
+
 
 //[auth]
 /// get a profile by it's id
@@ -37,7 +57,7 @@ router.get('/:id', async (req, res) => {
 
 /// create or update an existing profile
 // requires authentication
-router.put('/:id', async (req, res) => {
+router.put('/:id', auth, async (req, res) => {
 
     let profile = req.body;
 
@@ -57,6 +77,11 @@ router.put('/:id', async (req, res) => {
             //TODO: check if req.userID = oldProfile.userId
             //if no, kick the user with 403 and exit the function.
 
+            if (req.id_user !== oldProfile.id_user) {
+                res.status(403).send("You are not authorized to edit this profile.");
+                return;
+            }
+
             let result = db.updateProfile(profile);
         }
 
@@ -65,9 +90,10 @@ router.put('/:id', async (req, res) => {
 
             //I should check all the devices' device.id_user if they belong to the user trying to add the profile
 
+
             //TODO: verify and parse all the devices. Should prolly make a method in ./devices.js
             let oldDevice = await db.getDevice(device.id_configuration);
-            
+
             if (typeof oldDevice === 'undefined') {
                 let result = await db.addDevice(device); //TODO: figure out returning errors
             } else {
@@ -94,13 +120,14 @@ router.put('/:id', async (req, res) => {
 
 /// delete the specified profile
 // requires authentication
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', auth, async (req, res) => {
     const id = req.params.id;
 
-    //TODO: check if req.userID = oldProfile.userId
-    //if no, kick the user with 403.
-    //for this, I'll have to first fetch the profile from db :/
-
+    let oldProfile = await db.getProfile(id)
+    if (req.id_user !== oldProfile.id_user) {
+        res.status(403).send("You are not authorized to delete this profile.");
+        return;
+    }
 
     let result = db.deleteProfile(id); //TODO: error handling
     res.send("ok");
